// src/shared/api/generated/models/orgPaymentMethod.ts
/* /* eslint-disable *\/ */
/**
 * // @ts-nocheck
 */

export interface OrgPaymentMethod {
  color: string;
  created_at: string;
  display_order: number;
  icon: string;
  id: string;
  is_active: boolean;
  is_cash: boolean;
  label_translations: unknown;
  name: string;
  org_id: string;
  updated_at: string;
}